removed unneeded js load
