+++ GameSimulator/SASgame.html
@@ -26,7 +26,6 @@
 
   </div>
   <script src="js/papaparse.min.js"></script>
-  <script src="js/setSeeds.js"></script>
   <script src="js/SASgame.js"></script>
 
   <br>
