+++ GameSimulator/js/SASgame.js
@@ -157,6 +157,61 @@ var myGameArea = {
   },
 };
 
+// !---------------------- dynamic loading of csv data  -----------------------------!
+/**
+ * Array of setSeed Filenames
+ *
+ * @type {String[]}
+ */
+const setSeedFilenames = [
+  "setSeeds/1.csv",
+  "setSeeds/2.csv"
+];
+
+/**
+ * Load set seeds from file using papaparse.
+ * CSV loading is async, thus a callback function must
+ * be passed, which will be executed upon the load
+ * 
+ * @param {int} seedValue seed number, filename must be {seedValue}.csv
+ * @param {function} callback callback function to be executed onload
+ */
+function loadSetSeed(seedValue, callback) {
+
+  Papa.parse(setSeedFilenames[seedValue - 1], {
+    download: true,
+    complete: function (results) {
+      addCSVToSetSeeds(results, callback);
+    },
+    header: true,
+    dynamicTyping: true
+  });
+
+
+  function addCSVToSetSeeds(csv, callback) {
+    var readingRequests = false;
+    var PU = [];
+    var REQ = [];
+    var setSeed = [];
+
+    csv.data.forEach(function (row) {
+      if (row.startTime == null) { // empty line means we're now reading requests
+        readingRequests = true;
+        return; // continue
+      }
+      readingRequests ? REQ.push(row) : PU.push(row);
+
+
+    });
+    setSeed.PU = PU;
+    setSeed.REQ = REQ;
+    callback(setSeed);
+
+  }
+
+}
+
+
 // !---------------------- canvas event listeners  -----------------------------!
